backend/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1
Django==3.1
django-cors-headers==3.4.0
django-rest-knox==4.1.0
djangorestframework==3.11.1
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
pycparser==2.20
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
